control over smallcheck testing

+++ tests/test_examples.py
@@ -23,6 +23,8 @@ class TestExamples(TestCase):
 
         noTests = False
 
+        testSmallcheck = True
+
         skipTravis = [
             "redis",
             "simplejson",
@@ -211,25 +213,26 @@ class TestExamples(TestCase):
                                 rr2 = subprocess.call(["tstl_replay",
                                                        ".freesmall.test"])
                                 self.assertEqual(rr2, 255)
-                    scCmd = [
-                        "tstl_smallcheck",
-                        "--depth",
-                        "2",
-                        "--recursive",
-                        "1",
-                        "--multiple"]
-                    start = time.time()
-                    with open(os.devnull, 'w') as dnull:
-                        p = subprocess.Popen(scCmd)
-                    while (p.poll() is None) and ((time.time() - start) < 300):
-                        time.sleep(1)
-                    if p.poll() is None:
-                        p.terminate()
-                        print("TIMEOUT DURING SMALLCHECK!")
-                        timeoutFailures.append(f + "/" + t)
-                    else:
-                        r = p.returncode
-                        self.assertTrue(p in [0, 255])
+                    if testSmallcheck:
+                        scCmd = [
+                            "tstl_smallcheck",
+                            "--depth",
+                            "2",
+                            "--recursive",
+                            "1",
+                            "--multiple"]
+                        start = time.time()
+                        with open(os.devnull, 'w') as dnull:
+                            p = subprocess.Popen(scCmd)
+                        while (p.poll() is None) and ((time.time() - start) < 300):
+                            time.sleep(1)
+                        if p.poll() is None:
+                            p.terminate()
+                            print("TIMEOUT DURING SMALLCHECK!")
+                            timeoutFailures.append(f + "/" + t)
+                        else:
+                            r = p.returncode
+                            self.assertTrue(p in [0, 255])
                     os.remove("sut.py")
                     try:
                         os.remove("sut.pyc")
